+++ Source/VagabondsWork/VagabondsGameInstance.cpp
@@ -1,24 +1,12 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 #include "VagabondsGameInstance.h"
-#include "NavStaticBig.h"
-#include "Ship.h"
-#include "EngineUtils.h"
+#include "LevelActorsSubsystem.h"
 
 void UVagabondsGameInstance::OnStart()
 {
 	Super::OnStart();
-
-	RefreshTrackedActors();
-	if (UWorld* World = GetWorld())
-	{
-		World->GetTimerManager().SetTimer(
-			TrackedActorsRefreshTimer,
-			this,
-			&UVagabondsGameInstance::RefreshTrackedActors,
-			ActorListsRefreshInterval,
-			true);
-	}
+	GetSubsystem<ULevelActorsSubsystem>();
 }
 
 TArray<AActor*> UVagabondsGameInstance::GetActorsByFilter(const TArray<AActor*>& Actors, EActorFilterKind FilterKind, uint8 EnumValue) const
@@ -67,46 +55,3 @@ TArray<AActor*> UVagabondsGameInstance::GetActorsByFilter(const TArray<AActor*>&
 
 	return Result;
 }
-
-void UVagabondsGameInstance::RefreshTrackedActors()
-{
-	AllPlanets.Reset();
-	AllShips.Reset();
-
-	UWorld* World = GetWorld();
-	if (!World)
-	{
-		return;
-	}
-
-	for (TActorIterator<AActor> It(World); It; ++It)
-	{
-		AActor* Actor = *It;
-		if (!IsValid(Actor))
-		{
-			continue;
-		}
-
-		const UMarkerComponent* Marker = Actor->FindComponentByClass<UMarkerComponent>();
-		if (!IsValid(Marker))
-		{
-			continue;
-		}
-
-		if (Marker->MarkerType == EMarkerType::Planet)
-		{
-			if (ANavStaticBig* Planet = Cast<ANavStaticBig>(Actor))
-			{
-				AllPlanets.Add(Planet);
-			}
-		}
-		else if (Marker->MarkerType == EMarkerType::Ship)
-		{
-			if (AShip* ShipActor = Cast<AShip>(Actor))
-			{
-				AllShips.Add(ShipActor);
-			}
-		}
-	}
-}
-